2018 day 4 input and solutions
"ATM" is an introspective yet energetic track that highlights Cole's
relationship with money. He confirmed in a tweet that the song title is an
abbreviation for "Addicted To Money." He contemplates his current state of fame
and what it means for his family and his community. A money counter sound
effect is used throughout to drive home the monetary theme.

An audio recording of this track was leaked from his first-come-first-serve
concert on April 16, 2018, during which he banned cell phones and recording
devices.